requirements.txt
APScheduler==3.7.0
Django==3.2.4
djangorestframework==3.12.4
